Feature/patient full functionalities

* Provided Patient Height Model conversion, updated PatientBloodPressure page to get full list of blood pressures

* Implementation of PatientHeightsTable and PatientHeightsPage

* PatientHeightForm and Payload creation, creation of PatientMetricsFormContainer component

* Patient Heights get full list and add entry implementation

* Patient Height model append Doctor

* Initial implementation of PatientTemperaturesPage and PatientTemperaturesTable

* Append PatientTemperatures full list on load page

* Patient Temperature add form implementation

* Patient Weight Page and table implementation and usage

* Patient Weight Form implementation, add entry to weight records and load all records

* Provide formatted values for PatientBloodPressure

* Replace IntegerFormField with DoubleFormField, updated metrics payload to accept double values

* Localisation updates

* Formatting updates

* toLocalizedString implementation for unit enums

* Finalisation

* AppModule routing formatting update

* Provide Scaffold on PatientScaffold fallback to allow snackbar to display

+++ lib/models/patient_metrics/patient_blood_pressure_payload.dart
@@ -1,8 +1,8 @@
 import 'dart:convert';
 
 class PatientBloodPressurePayload {
-  int bloodPressureSystolic = 0;
-  int bloodPressureDiastolic = 0;
+  double bloodPressureSystolic = 0.0;
+  double bloodPressureDiastolic = 0.0;
 
   String toJson() {
     return jsonEncode({
